refactor(layout): fuse building and drawable structs
