+++ src/engine/layout.rs
@@ -1,11 +1,30 @@
-use crate::engine::test::{BuildingDrawable, RoadDrawable};
-use ansi_term::Color::{Blue, Green, Red};
-use serde::{Deserialize, Serialize};
+use std::fmt::Display;
 
-#[derive(Serialize, Deserialize, Debug)]
+use crate::ui::colors::{A_GREY_BLUE_COLOR, A_GREY_COLOR};
+use serde::{Deserialize, Serialize};
+use strum_macros::EnumString;
+
+use super::drawable::Drawable;
+
+#[derive(Debug, EnumString)]
+#[allow(non_camel_case_types)]
+pub enum BuildingType {
+    custom,
+    uniform,
+    empty_space,
+}
+
+impl Display for BuildingType {
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        write!(f, "{:?}", self)
+    }
+}
+
+#[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct Building {
     name: String,
     id: String,
+    district_id: usize,
     pos_x: i16,
     pos_y: i16,
     b_type: String,
@@ -15,26 +34,7 @@ pub struct Building {
     content: Option<Vec<String>>,
 }
 
-impl Building {
-    pub fn to_drawable(&self) -> BuildingDrawable {
-        BuildingDrawable {
-            xpos: self.pos_x,
-            ypos: self.pos_y,
-            color: if self.b_type == "empty_space" {
-                Green
-            } else {
-                Blue
-            },
-            content: self.content.clone(),
-            texture: self.texture,
-            width: self.width,
-            height: self.height,
-            b_type: self.b_type.clone(),
-        }
-    }
-}
-
-#[derive(Serialize, Deserialize, Debug)]
+#[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct Road {
     name: String,
     id: String,
@@ -44,20 +44,6 @@ pub struct Road {
     width: u8,
     length: u8,
     pavement: char,
-}
-
-impl Road {
-    pub fn to_drawable(&self) -> RoadDrawable {
-        RoadDrawable {
-            start_x: self.start_x,
-            start_y: self.start_y,
-            horizontal: self.horizontal,
-            pavement: self.pavement,
-            width: self.width,
-            length: self.length,
-            color: Red,
-        }
-    }
 }
 
 #[derive(Serialize, Deserialize, Debug)]
@@ -75,6 +61,14 @@ impl Layout {
         layout_obj
     }
 
+    pub fn load_core_layout() -> Self {
+        let layout = include_str!("../initial_data/starting_core.json");
+
+        let layout_obj: Layout = serde_json::from_str(layout).unwrap();
+
+        layout_obj
+    }
+
     pub fn add_building(&mut self, building: Building) {
         self.buildings.push(building);
     }
@@ -83,22 +77,151 @@ impl Layout {
         self.roads.push(road);
     }
 
-    pub fn get_buildings(&self) -> &Vec<Building> {
-        &self.buildings
+    /// Clone the vec
+    pub fn get_buildings(&self) -> Vec<Building> {
+        self.buildings.iter().map(|b| b.clone()).collect()
     }
 
-    pub fn get_roads(&self) -> &Vec<Road> {
-        &self.roads
+    /// Clone the vec
+    pub fn get_roads(&self) -> Vec<Road> {
+        self.roads.iter().map(|r| r.clone()).collect()
+    }
+}
+
+impl Drawable for Building {
+    fn x(&self) -> i16 {
+        self.pos_x
     }
 
-    pub fn get_road_drawables(&self) -> Vec<RoadDrawable> {
-        self.roads.iter().map(|road| road.to_drawable()).collect()
+    fn y(&self) -> i16 {
+        self.pos_y
     }
 
-    pub fn get_building_drawables(&self) -> Vec<BuildingDrawable> {
-        self.buildings
-            .iter()
-            .map(|building| building.to_drawable())
-            .collect()
+    fn width(&self) -> u8 {
+        if &self.b_type == "custom" {
+            let mut n: u8 = 0;
+            let lines = (self.content.as_ref()).unwrap();
+            for line in lines {
+                if n == 0 {
+                    n = line.chars().count() as u8
+                } else if (line.chars().count() as u8) != n {
+                    panic!("BuildingDrawable: Invalid line width, all lines need to have the same width")
+                }
+            }
+            n
+        } else {
+            self.width.unwrap()
+        }
+    }
+
+    fn height(&self) -> u8 {
+        if self.b_type == BuildingType::custom.to_string() {
+            return self.content.as_ref().unwrap().len() as u8;
+        }
+        self.height.unwrap()
+    }
+
+    fn shape(&self) -> String {
+        if self.b_type == BuildingType::custom.to_string() {
+            return self.content.as_ref().unwrap().join("\n");
+        } else if self.b_type == BuildingType::empty_space.to_string() {
+            if self.width.unwrap() == 1 && self.height.unwrap() == 1 {
+                return "▢\n".parse().unwrap();
+            } else if self.height.unwrap() == 1 {
+                let mut str: String = "╞".to_string();
+                str.push_str(&*"═".repeat(self.width.unwrap() as usize - 2));
+                str.push_str("╡\n");
+                return str;
+            } else if self.width.unwrap() == 1 {
+                let mut str: String = "╦\n".to_string();
+                str.push_str(&*"║\n".repeat(self.height.unwrap() as usize - 2));
+                str.push_str("╩\n");
+                return str;
+            } else {
+                let in_layers = self.height.unwrap() - 2;
+                let in_columns: u8 = self.width.unwrap() - 2;
+
+                let mut str: String = "╔".to_string();
+
+                //First Layer
+                str.push_str(&*"═".repeat(in_columns as usize));
+                str.push_str("╗\n");
+
+                for _ in 0..in_layers {
+                    let mut layer = "║".to_string();
+                    layer.push_str(&*" ".repeat(in_columns as usize));
+                    layer.push_str("║\n");
+                    str.push_str(&*layer);
+                }
+
+                str.push_str("╚");
+                str.push_str(&*"═".repeat(in_columns as usize));
+                str.push_str("╝\n");
+                return str;
+            }
+        }
+        let mut str: String = "".to_string();
+        for _ in 0..self.height.unwrap() {
+            str += &*(self
+                .texture
+                .unwrap()
+                .to_string()
+                .repeat(self.width.unwrap() as usize));
+            str += "\n";
+        }
+        str
+    }
+
+    fn color(&self) -> ansi_term::Color {
+        A_GREY_BLUE_COLOR
+    }
+}
+
+impl Drawable for Road {
+    fn x(&self) -> i16 {
+        self.start_x
+    }
+
+    fn y(&self) -> i16 {
+        self.start_y
+    }
+
+    fn width(&self) -> u8 {
+        if self.horizontal {
+            self.length
+        } else {
+            self.width
+        }
+    }
+
+    fn height(&self) -> u8 {
+        if self.horizontal {
+            self.width
+        } else {
+            self.length
+        }
+    }
+
+    fn shape(&self) -> String {
+        if self.horizontal {
+            let mut str = self.pavement.to_string().repeat(self.length as usize);
+            for _ in 0..self.width {
+                str += "\n";
+                str += &*self.pavement.to_string().repeat(self.length as usize)
+            }
+            str
+        } else {
+            let mut shape: String = self.pavement.to_string().repeat(self.width as usize);
+            for _ in 0..self.length {
+                shape += "\n";
+                shape += &*self.pavement.to_string().repeat(self.width as usize)
+            }
+
+            shape
+        }
+    }
+
+    fn color(&self) -> ansi_term::Color {
+        A_GREY_COLOR
     }
 }
