+++ src/engine/mod.rs
@@ -1,6 +1,5 @@
 pub mod drawable;
 pub mod viewport;
 pub mod core;
-pub mod test;
 pub mod keybinds;
 pub mod layout;
